Player actions and basic game flow
Each player can now take action on a betting round:
* Fold
* Check/call
* Bet/raise

The program now has a basic game flow, such as going to the next betting round after everyone has bet the same amount. However, win detectors and the end of a hand haven't been implemented yet.

Before an input, the state of the game is printed to the console, which are the player states, pot, and community cards. The user then inputs the action of each player onto the console.