Poker hand rankings
The new `HandRankings` class with a `calculate_ranking` method to determine the hand ranking for a set of cards (can be 5-7 cards). A test for hand rankings which shows 7 cards and the rank can now be run by calling the `hand_ranking_test` function in `testing.py`.

Other code changes:
* Ace (A) cards are now represented by the integer 14 instead of 1.
* "Hands" which represent a single deal are now referred as "Deals". This includes the renaming of the `Hand` class into `Deal`, and `current_hand` of the `PokerGame` class to `current_deal`.