Complete text based poker
The text based poker is now fully functional.

The program now runs a text based poker game just like before, but now with a win and lose system (hand ranking) and a more refined look.

After a player wins, the game advances to the next deal; and if there is only 1 player left, the entire game ends.