First Pygame testing
After finishing the rules engine, it's time to start developing the app GUI. Allin will use Pygame as its GUI engine.

The app is started by calling the `run` method of the `App` class in `app/app_main.py`. Just like most things in Pygame, the `App` class should be static and only act as a namespace (there could only be 1 instance of app).

The `App` class has one attribute for the current `Scene` class. A `Scene` class has a group of sprites and act as the parent of the current screen (game scene, main menu scene, etc). Inspired by the Unity game engine.

The button is first widget/game object is implemented in this Pygame development. A button is drawn by procedurally drawing a rounded rectangle with the pygame.draw module, and then drawing a text.