Table and cards GUI
The app now shows a poker table and 5 random cards.

Cards are drawn by drawing the base first from an image file, and then draw the suit and rank on top of it.

Upon starting the app, the starting scene is now the new `GameScene` class, which is a subclass of `Scene`.