Basic bet prompt structure
The bet prompt is used to ask the player how much they are raising/betting. The new bet prompt widget contains a bet slider, edit button, and confirm button. A new slider class is also created which is used as the parent class of the bet slider.

Both the button and slider classes now have a `global_rect` attribute, because if the button is part of another sprite, the x and y position of the rect will be relative to the parent sprite, while the mouse position is global (relative to the display surface).