Widget base class
The new `Widget` base class, along with the new positioning system of the `AutoRect` class, makes it much easier to position widgets. The `AutoRect` class is an extension of Pygame's Rect that has additional features: units (px or %), anchor, pivot, and parent.

Previously, widgets generally have `pos` and `dimensions` as their first two __init__ parameters. After this change, they are to be changed into `*rect_args`, so that non-keyword arguments are directly passed in as arguments for the `AutoRect` of a created widget.

Currently, the widgets that have been updated to use the new `Widget` class are the button, slider, and the bet prompt, along with all their components/subclasses.