Revamped widgets and its components
All the widgets have been updated to extend the `Widget` class instead of `pygame.sprite.Sprite`, as well as its components have been updated to use the new `WidgetComponent` class.

The constructor parameters of `Widget` and `WidgetComponent`, as well as most of its subclasses, have been updated to be (parent, *rect_args). A widget's parent could either be a scene or a widget, while a widget component's parent must be a widget. When creating a new widget/widget component, they are automatically added to the parent's main sprite group (all_sprites or children_group).

`Widget` and `WidgetComponent` share a same superclass, `AutoSprite`, an extension of Pygame's `Sprite` class.

`Widget` and `WidgetComponent` have a `move_anim` method, which simplifies code and removes redundant code that are responsible for moving objects.

Note that the revamping is still not fully done. There are still widget components whose positioning arguments can be simplified even further.