Settings data system
The new `settings_data.py` module enables storing the data of a collection of settings. The data can then be used by any part of the program, displayed in a setting panel, saved into an ".ini" file.

The settings data for the main app settings are put in the `global_settings.py` module, so that the program can access the data stored in the settings from any point.

Forgive for spaghetti code.