Pot initial restructure
The `pot` field in the `Deal` class is renamed to `pots` and is now a list instead of an integer. Other parts of the program has also been changed temporarily to suit the refactoring.

When someone bets on a round, the money now goes to the new `current_round_pot` field instead of `pot`.