Showdown with side pots
When there are side pots, the `showdown` method now checks for the winners of different side pots. After that, the winners are rewarded with the correct amount of money from the pot.

A new "refund" system where if a player goes all in but there is only that player on the new side pot, that side pot is then deleted and the money is returned to the all in player.