Thread waiter thing
The `ThreadWaiter` class is used by a coroutine to call a function, where the function call would be separated into a parallel thread, and the coroutine would wait for the result of that function.

Similar to the await keyword from Python, but I'm too skill issued to use it haiya.