Join room
Sending a "join <code>" basic request to the server now causes the client to join a room with the given code.

The `AllinServer` class now has a `rooms` attribute, a dict containing all the active rooms (`ServerGameRoom` instances) in the server, with the `ServerGameRoom` class being the subclass of `PokerGame`.