Leave room and game event packet
When a player leaves a game that isn't ongoing, they immediately get removed from the player list.

When a client gets forcefully disconnected from the server, the scene turns back into the main menu.

When the `HandlerPlayer` class receives a game event from the `ServerGameRoom` broadcast, it sends a game event packet to its client, in which the client receives that packet and do some stuff to the `MultiplayerGame` (WIP).