Multiplayer reset players
Joining the test room on the multiplayer menu now actually shows all the players that are currently in the room, with the server sending a game data packet containing the list of players, and then sending a reset players game event.

When the client receives a game event or game data packet, it appends it to a game event queue. That queue is then read by the `MultiplayerGame`, which is updated every tick.