Index-based winners list
The `winners` list of the `Hand` class is now a 2D list of integers instead of a 2D list of `PlayerHand` objects. This makes sending the winners list via the game data packets much easier. Thus, the showdown mechanism now works properly on multiplayer games.

Additionally, players who are eliminated in `ServerGameRoom` are now moved to the `spectators` list instead of being deleted. The server game now also broadcasts game events to those players in the `spectators` list.