Join and leave mid-game
Players can now join a room when a game is in progress. Doing so would put the player in the joining queue, and when a new hand starts, the players in the queue would then join the game.

When a player disconnects or leaves a room mid-game, their player instance would still be in the room until the next hand starts. The server would automatically play their next action, which is to fold.